requirements.txt
fastapi==0.78.0
numpy==1.22.4
Pillow==9.1.1
tensorflow==2.9.1
uvicorn==0.17.6
python-multipart==0.0.5
gunicorn==20.1.0
